refactor: 거대한 테스트 클래스를 쪼개기
